Refactor SqlIndexDataStore
