+++ src/Microsoft.Health.Dicom.SqlServer/Features/Store/SqlIndexDataStoreV1.cs
@@ -1,0 +1,275 @@
+﻿// -------------------------------------------------------------------------------------------------
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
+// -------------------------------------------------------------------------------------------------
+
+using System;
+using System.Collections.Generic;
+using System.Data;
+using System.Threading;
+using System.Threading.Tasks;
+using Dicom;
+using EnsureThat;
+using Microsoft.Data.SqlClient;
+using Microsoft.Health.Dicom.Core.Exceptions;
+using Microsoft.Health.Dicom.Core.Extensions;
+using Microsoft.Health.Dicom.Core.Features.CustomTag;
+using Microsoft.Health.Dicom.Core.Features.Model;
+using Microsoft.Health.Dicom.Core.Features.Store;
+using Microsoft.Health.Dicom.Core.Models;
+using Microsoft.Health.Dicom.SqlServer.Features.Schema.Model;
+using Microsoft.Health.SqlServer.Features.Client;
+using Microsoft.Health.SqlServer.Features.Storage;
+
+namespace Microsoft.Health.Dicom.SqlServer.Features.Store
+{
+    internal class SqlIndexDataStoreV1 : IIndexDataStore
+    {
+        private readonly SqlConnectionWrapperFactory _sqlConnectionFactoryWrapper;
+
+        public SqlIndexDataStoreV1(
+            SqlConnectionWrapperFactory sqlConnectionWrapperFactory)
+        {
+            EnsureArg.IsNotNull(sqlConnectionWrapperFactory, nameof(sqlConnectionWrapperFactory));
+            _sqlConnectionFactoryWrapper = sqlConnectionWrapperFactory;
+        }
+
+        public virtual async Task<long> CreateInstanceIndexAsync(DicomDataset instance, IEnumerable<IndexTag> indexableDicomTags, CancellationToken cancellationToken)
+        {
+            EnsureArg.IsNotNull(instance, nameof(instance));
+            EnsureArg.IsNotNull(indexableDicomTags, nameof(indexableDicomTags));
+
+            using (SqlConnectionWrapper sqlConnectionWrapper = await _sqlConnectionFactoryWrapper.ObtainSqlConnectionWrapperAsync(cancellationToken))
+            using (SqlCommandWrapper sqlCommandWrapper = sqlConnectionWrapper.CreateSqlCommand())
+            {
+                V1.AddInstance.PopulateCommand(
+                      sqlCommandWrapper,
+                      instance.GetString(DicomTag.StudyInstanceUID),
+                      instance.GetString(DicomTag.SeriesInstanceUID),
+                      instance.GetString(DicomTag.SOPInstanceUID),
+                      instance.GetSingleValueOrDefault<string>(DicomTag.PatientID),
+                      instance.GetSingleValueOrDefault<string>(DicomTag.PatientName),
+                      instance.GetSingleValueOrDefault<string>(DicomTag.ReferringPhysicianName),
+                      instance.GetStringDateAsDateTime(DicomTag.StudyDate),
+                      instance.GetSingleValueOrDefault<string>(DicomTag.StudyDescription),
+                      instance.GetSingleValueOrDefault<string>(DicomTag.AccessionNumber),
+                      instance.GetSingleValueOrDefault<string>(DicomTag.Modality),
+                      instance.GetStringDateAsDateTime(DicomTag.PerformedProcedureStepStartDate),
+                      (byte)IndexStatus.Creating);
+
+                try
+                {
+                    return (long)(await sqlCommandWrapper.ExecuteScalarAsync(cancellationToken));
+                }
+                catch (SqlException ex)
+                {
+                    switch (ex.Number)
+                    {
+                        case SqlErrorCodes.Conflict:
+                            {
+                                if (ex.State == (byte)IndexStatus.Creating)
+                                {
+                                    throw new PendingInstanceException();
+                                }
+
+                                throw new InstanceAlreadyExistsException();
+                            }
+                    }
+
+                    throw new DataStoreException(ex);
+                }
+            }
+        }
+
+        public async Task DeleteInstanceIndexAsync(string studyInstanceUid, string seriesInstanceUid, string sopInstanceUid, DateTimeOffset cleanupAfter, CancellationToken cancellationToken)
+        {
+            EnsureArg.IsNotNullOrEmpty(studyInstanceUid, nameof(studyInstanceUid));
+            EnsureArg.IsNotNullOrEmpty(seriesInstanceUid, nameof(seriesInstanceUid));
+            EnsureArg.IsNotNullOrEmpty(sopInstanceUid, nameof(sopInstanceUid));
+
+            await DeleteInstanceAsync(studyInstanceUid, seriesInstanceUid, sopInstanceUid, cleanupAfter, cancellationToken);
+        }
+
+        public async Task DeleteSeriesIndexAsync(string studyInstanceUid, string seriesInstanceUid, DateTimeOffset cleanupAfter, CancellationToken cancellationToken)
+        {
+            EnsureArg.IsNotNullOrEmpty(studyInstanceUid, nameof(studyInstanceUid));
+            EnsureArg.IsNotNullOrEmpty(seriesInstanceUid, nameof(seriesInstanceUid));
+
+            await DeleteInstanceAsync(studyInstanceUid, seriesInstanceUid, sopInstanceUid: null, cleanupAfter, cancellationToken);
+        }
+
+        public async Task DeleteStudyIndexAsync(string studyInstanceUid, DateTimeOffset cleanupAfter, CancellationToken cancellationToken)
+        {
+            EnsureArg.IsNotNullOrEmpty(studyInstanceUid, nameof(studyInstanceUid));
+
+            await DeleteInstanceAsync(studyInstanceUid, seriesInstanceUid: null, sopInstanceUid: null, cleanupAfter, cancellationToken);
+        }
+
+        public async Task UpdateInstanceIndexStatusAsync(
+            VersionedInstanceIdentifier versionedInstanceIdentifier,
+            IndexStatus status,
+            CancellationToken cancellationToken)
+        {
+            EnsureArg.IsNotNull(versionedInstanceIdentifier, nameof(versionedInstanceIdentifier));
+            EnsureArg.IsTrue(Enum.IsDefined(typeof(IndexStatus), status));
+            EnsureArg.IsTrue((int)status < byte.MaxValue);
+
+            using (SqlConnectionWrapper sqlConnectionWrapper = await _sqlConnectionFactoryWrapper.ObtainSqlConnectionWrapperAsync(cancellationToken))
+            using (SqlCommandWrapper sqlCommandWrapper = sqlConnectionWrapper.CreateSqlCommand())
+            {
+                VLatest.UpdateInstanceStatus.PopulateCommand(
+                    sqlCommandWrapper,
+                    versionedInstanceIdentifier.StudyInstanceUid,
+                    versionedInstanceIdentifier.SeriesInstanceUid,
+                    versionedInstanceIdentifier.SopInstanceUid,
+                    versionedInstanceIdentifier.Version,
+                    (byte)status);
+
+                try
+                {
+                    await sqlCommandWrapper.ExecuteScalarAsync(cancellationToken);
+                }
+                catch (SqlException ex)
+                {
+                    switch (ex.Number)
+                    {
+                        case SqlErrorCodes.NotFound:
+                            throw new InstanceNotFoundException();
+
+                        default:
+                            throw new DataStoreException(ex);
+                    }
+                }
+            }
+        }
+
+        public async Task<IEnumerable<VersionedInstanceIdentifier>> RetrieveDeletedInstancesAsync(int batchSize, int maxRetries, CancellationToken cancellationToken = default)
+        {
+            var results = new List<VersionedInstanceIdentifier>();
+
+            using (SqlConnectionWrapper sqlConnectionWrapper = await _sqlConnectionFactoryWrapper.ObtainSqlConnectionWrapperAsync(cancellationToken, true))
+            using (SqlCommandWrapper sqlCommandWrapper = sqlConnectionWrapper.CreateSqlCommand())
+            {
+                VLatest.RetrieveDeletedInstance.PopulateCommand(
+                    sqlCommandWrapper,
+                    batchSize,
+                    maxRetries);
+
+                using (var reader = await sqlCommandWrapper.ExecuteReaderAsync(CommandBehavior.SequentialAccess, cancellationToken))
+                {
+                    try
+                    {
+                        while (await reader.ReadAsync(cancellationToken))
+                        {
+                            (string studyInstanceUid, string seriesInstanceUid, string sopInstanceUid, long watermark) = reader.ReadRow(
+                                VLatest.DeletedInstance.StudyInstanceUid,
+                                VLatest.DeletedInstance.SeriesInstanceUid,
+                                VLatest.DeletedInstance.SopInstanceUid,
+                                VLatest.DeletedInstance.Watermark);
+
+                            results.Add(new VersionedInstanceIdentifier(
+                                studyInstanceUid,
+                                seriesInstanceUid,
+                                sopInstanceUid,
+                                watermark));
+                        }
+                    }
+                    catch (SqlException ex)
+                    {
+                        throw new DataStoreException(ex);
+                    }
+                }
+            }
+
+            return results;
+        }
+
+        public async Task DeleteDeletedInstanceAsync(VersionedInstanceIdentifier versionedInstanceIdentifier, CancellationToken cancellationToken = default)
+        {
+            using (SqlConnectionWrapper sqlConnectionWrapper = await _sqlConnectionFactoryWrapper.ObtainSqlConnectionWrapperAsync(cancellationToken, true))
+            using (SqlCommandWrapper sqlCommandWrapper = sqlConnectionWrapper.CreateSqlCommand())
+            {
+                VLatest.DeleteDeletedInstance.PopulateCommand(
+                    sqlCommandWrapper,
+                    versionedInstanceIdentifier.StudyInstanceUid,
+                    versionedInstanceIdentifier.SeriesInstanceUid,
+                    versionedInstanceIdentifier.SopInstanceUid,
+                    versionedInstanceIdentifier.Version);
+
+                try
+                {
+                    await sqlCommandWrapper.ExecuteScalarAsync(cancellationToken);
+                }
+                catch (SqlException ex)
+                {
+                    throw new DataStoreException(ex);
+                }
+            }
+        }
+
+        public async Task<int> IncrementDeletedInstanceRetryAsync(VersionedInstanceIdentifier versionedInstanceIdentifier, DateTimeOffset cleanupAfter, CancellationToken cancellationToken = default)
+        {
+            using (SqlConnectionWrapper sqlConnectionWrapper = await _sqlConnectionFactoryWrapper.ObtainSqlConnectionWrapperAsync(cancellationToken, true))
+            using (SqlCommandWrapper sqlCommandWrapper = sqlConnectionWrapper.CreateSqlCommand())
+            {
+                VLatest.IncrementDeletedInstanceRetry.PopulateCommand(
+                    sqlCommandWrapper,
+                    versionedInstanceIdentifier.StudyInstanceUid,
+                    versionedInstanceIdentifier.SeriesInstanceUid,
+                    versionedInstanceIdentifier.SopInstanceUid,
+                    versionedInstanceIdentifier.Version,
+                    cleanupAfter);
+
+                try
+                {
+                    return (int)(await sqlCommandWrapper.ExecuteScalarAsync(cancellationToken));
+                }
+                catch (SqlException ex)
+                {
+                    throw new DataStoreException(ex);
+                }
+            }
+        }
+
+        private async Task DeleteInstanceAsync(string studyInstanceUid, string seriesInstanceUid, string sopInstanceUid, DateTimeOffset cleanupAfter, CancellationToken cancellationToken)
+        {
+            using (SqlConnectionWrapper sqlConnectionWrapper = await _sqlConnectionFactoryWrapper.ObtainSqlConnectionWrapperAsync(cancellationToken))
+            using (SqlCommandWrapper sqlCommandWrapper = sqlConnectionWrapper.CreateSqlCommand())
+            {
+                VLatest.DeleteInstance.PopulateCommand(
+                    sqlCommandWrapper,
+                    cleanupAfter,
+                    (byte)IndexStatus.Created,
+                    studyInstanceUid,
+                    seriesInstanceUid,
+                    sopInstanceUid);
+
+                try
+                {
+                    await sqlCommandWrapper.ExecuteScalarAsync(cancellationToken);
+                }
+                catch (SqlException ex)
+                {
+                    switch (ex.Number)
+                    {
+                        case SqlErrorCodes.NotFound:
+                            if (!string.IsNullOrEmpty(sopInstanceUid))
+                            {
+                                throw new InstanceNotFoundException();
+                            }
+
+                            if (!string.IsNullOrEmpty(seriesInstanceUid))
+                            {
+                                throw new SeriesNotFoundException();
+                            }
+
+                            throw new StudyNotFoundException();
+
+                        default:
+                            throw new DataStoreException(ex);
+                    }
+                }
+            }
+        }
+    }
+}
